Fixed menu retrieving Fixed price datatype Fixed bug in Menu initialisation Cleaned up a small amount of code

+++ src/main/java/sql/Foods.java
@@ -52,7 +52,7 @@ public class Foods {
                     resultSet.getString("food_name"),
                     resultSet.getString("food_description"),
                     resultSet.getInt("calories"),
-                    resultSet.getLong("price"),
+                    resultSet.getBigDecimal("price"),
                     resultSet.getBoolean("available"),
                     resultSet.getInt("category_id"));
         }
